Implement a Camera
Implement a Camera struct that acts like a virtual camera taking
pictures of a scene.

The camera maps a three-dimensional scene onto a two-dimensional canvas.

The camera's canvas is always exactly one unit in front on the camera.

The ray_for_pixel() function constructs a ray that passes through a
point in world space given x & y canvas coordinates

The render() function casts a ray through each of the pixels of the
canvas and writes the shaded color value to the canvas.